add matching start component with backup
